--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "terminal.integrated.scrollback": 2000
-}
