+++ requirements.txt
@@ -1,64 +1,9 @@
-absl-py==2.1.0
-asttokens==2.4.1
-astunparse==1.6.3
-certifi==2024.8.30
-charset-normalizer==3.3.2
-contourpy==1.3.0
-cycler==0.12.1
-decorator==5.1.1
-executing==2.1.0
-flatbuffers==24.3.25
-fonttools==4.53.1
-gast==0.6.0
-google-pasta==0.2.0
-grpcio==1.66.1
-h5py==3.11.0
-idna==3.10
-ipython==8.27.0
-jedi==0.19.1
 keras==3.5.0
-kiwisolver==1.4.7
-libclang==18.1.1
-Markdown==3.7
-markdown-it-py==3.0.0
-MarkupSafe==2.1.5
 matplotlib==3.9.2
-matplotlib-inline==0.1.7
-mdurl==0.1.2
-mediapy==1.2.2
-ml-dtypes==0.4.1
-namex==0.0.8
-numpy==1.26.4
-opencv-python==4.10.0.84
-opt-einsum==3.3.0
-optree==0.12.1
-packaging==24.1
-parso==0.8.4
-pexpect==4.9.0
-pillow==10.4.0
-prompt_toolkit==3.0.47
-protobuf==4.25.5
-ptyprocess==0.7.0
-pure_eval==0.2.3
-Pygments==2.18.0
-pyparsing==3.1.4
-python-dateutil==2.9.0.post0
-requests==2.32.3
-rich==13.8.1
-setuptools==75.1.0
-six==1.16.0
-stack-data==0.6.3
-tensorboard==2.17.1
-tensorboard-data-server==0.7.2
+numpy==2.1.3
+opencv_python==4.10.0.84
+pandas==2.2.3
+scikit_learn==1.5.2
+tabulate==0.9.0
 tensorflow==2.17.0
-tensorflow-hub==0.16.1
-termcolor==2.4.0
-tf_keras==2.17.0
-tqdm==4.66.5
-traitlets==5.14.3
-typing_extensions==4.12.2
-urllib3==2.2.3
-wcwidth==0.2.13
-Werkzeug==3.0.4
-wheel==0.44.0
-wrapt==1.16.0
+tensorflow_intel==2.17.0
